api/requirements.txt
# FastAPI and Uvicorn for API
fastapi==0.95.1
uvicorn==0.22.0

# OCR service integration
requests==2.31.0

# Google API and OAuth libraries
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.102.0

# Optional: For managing Python environment if necessary
python-dotenv==1.0.0
